+++ utils/test/connection.go
@@ -9,8 +9,6 @@ package test
 import (
 	"context"
 	"crypto/tls"
-	"fmt"
-	"path/filepath"
 	"strconv"
 	"testing"
 	"time"
@@ -18,8 +16,6 @@ import (
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
-
-	"github.com/hyperledger/fabric-x-common/tools/cryptogen"
 
 	"github.com/hyperledger/fabric-x-common/utils/connection"
 	"github.com/hyperledger/fabric-x-common/utils/retry"
@@ -31,14 +27,6 @@ var (
 	InsecureTLSConfig connection.TLSConfig
 	// defaultGrpcRetryProfile defines the retry policy for a gRPC client connection.
 	defaultGrpcRetryProfile retry.Profile
-
-	// OrgRootCA is the path to organization 0's TLS client credentials in the crypto materials directory.
-	OrgRootCA = filepath.Join(cryptogen.PeerOrganizationsDir, "peer-org-0.com",
-		cryptogen.MSPDir, cryptogen.TLSCaCertsDir, "tlspeer-org-0-CA-cert.pem")
-
-	// OrdererRootCATLSPath is the path to organization 0's orderer TLS credentials in the crypto materials directory.
-	OrdererRootCATLSPath = filepath.Join(cryptogen.OrdererOrganizationsDir,
-		"orderer-org-0.com", cryptogen.MSPDir, cryptogen.TLSCaCertsDir, "tlsorderer-org-0-CA-cert.pem")
 )
 
 // CheckServerStopped returns true if the grpc server listening on a
@@ -202,21 +190,6 @@ func NewPreAllocatedLocalHostServerConfig(t *testing.T, tlsConfig connection.TLS
 	return serverConfig
 }
 
-// NewServiceTLSConfig creates a server TLS configuration with certificates loaded from the artifact path.
-// This function constructs paths to TLS certificates for a given service within the peer organization structure.
-func NewServiceTLSConfig(artifactsPath, serviceName, mode string) connection.TLSConfig {
-	subPath := filepath.Join(artifactsPath, cryptogen.PeerOrganizationsDir, "peer-org-0.com",
-		cryptogen.PeerNodesDir, serviceName, cryptogen.TLSDir)
-	return connection.TLSConfig{
-		Mode:     mode,
-		CertPath: filepath.Join(subPath, "server.crt"),
-		KeyPath:  filepath.Join(subPath, "server.key"),
-		CACertPaths: []string{
-			filepath.Join(artifactsPath, OrgRootCA),
-		},
-	}
-}
-
 // NewLocalHostServiceConfig returns a grpcservice.ServerConfig with both gRPC and monitoring endpoints.
 // Both endpoints use "localhost:0" (auto-assigned ports) with the given TLS credentials.
 func NewLocalHostServiceConfig(creds connection.TLSConfig) *serve.Config {
@@ -240,20 +213,5 @@ func NewLocalHostServer(creds connection.TLSConfig) *serve.ServerConfig {
 	return &serve.ServerConfig{
 		Endpoint: connection.Endpoint{Host: "127.0.0.1"},
 		TLS:      creds,
-	}
-}
-
-// OrgClientTLSConfig creates a mutual TLS client configuration using a specific
-// peer organization's TLS client certificate. The serverCACertPaths are the CA
-// certs needed to verify the server (typically from the CredentialsFactory).
-func OrgClientTLSConfig(artifactsPath string, orgIndex int, serverCACertPaths []string) connection.TLSConfig {
-	orgDomain := fmt.Sprintf("peer-org-%d.com", orgIndex)
-	tlsDir := filepath.Join(artifactsPath, cryptogen.PeerOrganizationsDir, orgDomain,
-		cryptogen.UsersDir, fmt.Sprintf("client@%s", orgDomain), cryptogen.TLSDir)
-	return connection.TLSConfig{
-		Mode:        connection.MutualTLSMode,
-		CertPath:    filepath.Join(tlsDir, "client.crt"),
-		KeyPath:     filepath.Join(tlsDir, "client.key"),
-		CACertPaths: serverCACertPaths,
 	}
 }
